# go-backend/internal/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: TRPG Game Public API
  version: 2.3.0-m2.3-b
  license:
    name: Project license not specified
  description: |
    TRPG Game 对前端开放的 Go REST API，包含 M2.1 好友关系、M2.2 私聊和 M2.3-B 群组持久化接口。

    所有业务响应均使用 `code` 和 `message`；除删除剧本外，成功响应还包含 `data`。
    `code = 0` 表示业务成功。WebSocket、Python AI 和 Go 内部回调不属于本公共规范。
servers:
  - url: /
    description: 当前 Swagger UI 所在服务
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/games/solo/start:
    post:
      tags: [Games]
      summary: 开始单人游戏
      description: 剧本必须为 ready，角色必须属于该剧本。
      operationId: startSoloGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSoloGameRequest'
            example: {script_id: 9, character_id: 12}
      responses:
        '201':
          description: 游戏创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartSoloGameResponse'
        '400':
          description: 请求无效（业务码 1300）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1300, message: invalid game request}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 剧本不存在（1301）或角色不存在（1303）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                scriptNotFound:
                  value: {code: 1301, message: script not found}
                characterNotFound:
                  value: {code: 1303, message: character not found in script}
        '409':
          description: 剧本未就绪（1302）或游戏创建冲突（1305）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              examples:
                scriptNotReady:
                  value: {code: 1302, message: script is not ready}
                startConflict:
                  value: {code: 1305, message: game start status conflict}
        '503':
          description: AI 开场叙事不可用（业务码 1304）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1304, message: AI opening generation unavailable}
        '500':
          description: 内部错误（业务码 1306）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1306, message: internal error}
...
  /api/v1/games/{roomId}/skip:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      tags: [Games]
      summary: 当前行动者主动跳过多人回合
      operationId: skipMultiplayerTurn
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SkipTurnRequest'}
      responses:
        '200':
          description: 回合已跳过或幂等重放
          content:
            application/json:
              schema: {$ref: '#/components/schemas/SkipTurnResponse'}
        '400':
          description: 请求无效（1310）
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403':
          description: 不是当前行动者（1922）
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
        '404':
          description: 房间或成员不存在（1311）
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
        '409':
          description: 回合冲突、已有行动或请求 ID 冲突（1921/1923/1924）
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
        '503':
          description: 多人运行态不可用（1920）
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ApiError'}
  /api/v1/games/{roomId}/save:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      tags: [Games]
      summary: 创建手动存档
      operationId: createManualSave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualSaveRequest'
            example: {save_name: 进入密室前}
      responses:
        '201':
          description: 存档创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManualSaveResponse'
        '400':
          description: 请求无效（业务码 1321）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1321, message: invalid game save request}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/GameRoomNotFound'
        '409':
          description: 当前房间状态不可存档（业务码 1322）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1322, message: game room cannot be saved}
        '503':
          $ref: '#/components/responses/GameRuntimeUnavailable'
        '500':
          description: 内部错误（业务码 1323）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1323, message: internal error}
  /api/v1/games/{roomId}/saves:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    get:
      tags: [Games]
      summary: 查询房间存档列表
      operationId: listGameSaves
      responses:
        '200':
          description: 查询成功；不返回 Redis 快照和消息正文
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameSaveListResponse'
        '400':
          description: 请求无效（业务码 1324）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1324, message: invalid game save query}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/GameRoomNotFound'
        '500':
          description: 内部错误（业务码 1325）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1325, message: internal error}
  /api/v1/games/{roomId}/load:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      tags: [Games]
      summary: 从存档恢复游戏
      description: 恢复完成后房间保持 paused，需要再调用恢复接口。
      operationId: loadGame
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadGameRequest'
            example: {save_id: 91}
      responses:
        '200':
          description: 读档成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadGameResponse'
        '400':
          description: 请求无效（业务码 1332）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1332, message: invalid game load request}
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 房间不存在（1311）或存档不存在（1333）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: 存档损坏（1334）或当前房间状态不可读档（1335）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          $ref: '#/components/responses/GameRuntimeUnavailable'
        '500':
          description: 内部错误（业务码 1336）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example: {code: 1336, message: internal error}
  /api/v1/games/{roomId}/pause:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      tags: [Games]
      summary: 暂停游戏
      operationId: pauseGame
      responses:
        '200':
          description: 暂停成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusResponse'
              example:
                code: 0
                message: ok
                data: {room_id: 41, status: paused}
        '400':
          description: 请求无效（业务码 1326）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/GameRoomNotFound'
        '409':
          description: 当前状态不可暂停（业务码 1327）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          $ref: '#/components/responses/GameRuntimeUnavailable'
        '500':
          description: 内部错误（业务码 1328）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RoomId:
      name: roomId
      in: path
      required: true
      description: 游戏房间 ID；单人游戏接口限房主，多人大厅接口限有效成员
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 41
...
  responses:
    Unauthorized:
      description: 缺少 JWT（1001）或 JWT 无效/过期（1002）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            missingToken:
              value: {code: 1001, message: Missing authorization header}
            invalidToken:
              value: {code: 1002, message: Invalid or expired token}
...
    GameRoomNotFound:
      description: 房间不存在或不属于当前用户（业务码 1311）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example: {code: 1311, message: game room not found}
...
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          format: int32
          minimum: 1
        message:
          type: string
...
    StartSoloGameRequest:
      type: object
      required: [script_id, character_id]
      properties:
        script_id:
          type: integer
          format: int64
          minimum: 1
        character_id:
          type: integer
          format: int64
          minimum: 1
...
    StartSoloGameResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/StartSoloGameData'
...
    SkipTurnRequest:
      type: object
      required: [request_id, expected_turn]
      properties:
        request_id: {type: string, format: uuid}
        expected_turn: {type: integer, minimum: 0}
...
    SkipTurnResponse:
      type: object
      required: [code, message, data]
      properties:
        code: {type: integer, enum: [0]}
        message: {type: string, example: ok}
        data: {$ref: '#/components/schemas/SkipTurnData'}
    ManualSaveRequest:
      type: object
      required: [save_name]
      properties:
        save_name:
          type: string
          minLength: 1
          maxLength: 256
          description: 去除首尾空白后不能为空，且不能包含控制字符
...
    ManualSaveResponse:
      type: object
      required: [code, message, data]
      properties:
        code:
          type: integer
          enum: [0]
        message:
          type: string
          example: ok
        data:
          $ref: '#/components/schemas/ManualSaveData'
...